[feat] making json support a feature, adding support to make plain text requests.
