+++ Cargo.toml
@@ -10,7 +10,8 @@ license = "GPL-3.0-or-later"
 keywords = ["http", "networking", "client", "fetch", "rest-client"]
 
 [features]
-default = ["tokio-rt"]
+default = ["tokio-rt", "json"]
+json = ["dep:serde_json", "dep:serde"]
 tokio-rt = []
 smol-rt = ["dep:smol", "dep:smol-hyper", "dep:async-native-tls", "dep:smol-macros", "dep:macro_rules_attribute"]
 compio-rt = ["dep:compio", "dep:cyper-core", "dep:async-native-tls"]
@@ -33,8 +34,8 @@ hyper-util = { version = "0.1.10", features = [
   "tokio",
 ] }
 macro_rules_attribute = { version = "0.2.2", optional = true }
-serde = { version = "1.0.217", features = ["derive"] }
-serde_json = "1.0.138"
+serde = { version = "1.0.217", features = ["derive"], optional = true }
+serde_json = { version = "1.0.138", optional = true }
 smol = { version = "2.0.2", optional = true }
 smol-hyper = { version = "0.1.0", optional = true }
 smol-macros = { version = "0.1.1", optional = true }
